Add cc65 tools for 6502 assembly support
This commit introduces the cc65 toolchain into the project by updating
the `default.nix` file. The inclusion of cc65 allows developers to work
with 6502 assembly language, expanding the capabilities of the
environment for retro computing projects or anyone interested in 8-bit
development. The change ensures that the necessary tools are readily
available for users, aligning with the project's goal of supporting a
broader range of programming languages and tools.