Remove no longer used vasm compiler
This commit removes the vasm compiler from the project, which is no
longer utilized. The `vasm.nix` file has been deleted, and references to
vasm have been removed from the `default.nix` file. This cleanup
enhances the maintainability of the codebase by eliminating obsolete
dependencies, thereby streamlining the build process for users. The
decision aligns with the project's ongoing effort to maintain a lean and
efficient development environment.