Fix minipro derivation
Updated the minipro package in response to recent changes and
improvements. The version has been bumped from 0.5 to 0.7.2, and the
source revision has been updated to reflect the latest commit from the
GitLab repository. Additionally, the native build inputs were adjusted
to include `zlib`, and the build inputs now also include `libb64` to
enhance functionality. The changes in `default.nix` re-enabled the
minipro package in the home environment setup, ensuring it is properly
included alongside the `dcc6502` disassembler. These modifications align
with the need to ensure the minipro utility is updated and fully
functioning for users requiring EEPROM programming capabilities.