Add LibreOffice to edit/view XLSX files
This commit introduces LibreOffice to the default Nix GUI module to
provide users with the ability to view and edit XLSX files. By including
LibreOffice in the configuration, we enhance the suite of applications
available for managing spreadsheet files, thereby improving the overall
usability of the GUI environment. This change allows users to seamlessly
work with Excel documents, facilitating better productivity and
compatibility with common file formats.

The key modification is the addition of the `libreoffice` line to the
list within the `modules/gui/default.nix` configuration file.