Trigger branch initialization on issue-description change
This commit adds functionality to automatically initialize a feature
branch when the issue description buffer is closed. By introducing an
autocommand that triggers the `gitInitFeatureBranch` function upon
unloading the `/tmp/issue-description` file, we facilitate a smoother
workflow for developers. This enhancement minimizes manual steps and
allows for immediate setup of branches based on user stories, further
streamlining the development process.

Key changes include the removal of the mapping in the `fugitive.nix`
file that previously invoked branch initialization manually, as well as
the addition of a new autocommand in `autocmd.vim` that automatically
triggers the feature branch initialization when the issue description is
no longer in the buffer.