Update font repository reference in `fonts.nix`
Updated the commit reference for the custom fonts repository from an
outdated revision to the latest one to ensure the application uses the
most current version of the fonts. This change is necessary for
maintaining consistency and compatibility with font updates, which
directly impacts the visual presentation of the GUI. The updated
revision number is `5df129be8ff5c805334656a95e7adc4953d0e35a`.

This commit helps address the user story requirement regarding the use
of updated font resources in the application.