Add trusted-users configuration to nix settings
This commit introduces a new line in the `configuration.nix` file to set
`trusted-users` to include `root` and `@wheel`. This change is made to
ensure that users in the `wheel` group are granted the necessary
permissions for Nix operations, enhancing security and usability for
managing Nix packages on the Omega machine. By implementing this
configuration, we align with the user story requirements regarding user
permissions and system management.