Disable freetube module import
Commented out the import of the `freetube.nix` module in `default.nix`
to prevent it from being included in the build. This adjustment was made
to address issues related to application conflicts and ensure a smoother
development environment when working with other modules like
`tertius.nix`, `nnn.nix`, and `clipboard.nix`. This change improves
modularity and reduces potential errors during runtime.