Improve SSH public key formatting in configuration.nix
This commit updates the formatting of the SSH public key for the 'git'
user in the `configuration.nix` file. The previous version had the
public key split across multiple lines, which could lead to issues when
attempting to authenticate. By consolidating the key into a single line,
we ensure that it is correctly recognized and functionally usable
without losing any part of it, thereby improving the reliability of user
connections via SSH.