Update GPG agent configuration in Nix file
This commit modifies the GPG agent configuration in `default.nix` by
commenting out the line that specifies `pinentryPackage`. This change
aligns with the user story requirement to enable flexibility in GPG
pinentry options by allowing users to configure their preferred pinentry
package in subsequent modifications. Additionally, it maintains
compatibility with various environments that might not support the GTK
version of the pinentry package.