Enable pinentry package in GPG agent configuration
Updated the GPG agent configuration to uncomment the `pinentryPackage`
line, enabling the use of `pkgs.pinentry-gtk2`. This change was made to
fulfill the requirement of using a specific pinentry package, which is
necessary for the GPG agent to handle passphrase prompts more
effectively in graphical environments. This modification improves user
experience during secure operations involving GPG.