Add Traefik group configuration to Nix configuration
This commit modifies the Traefik service configuration in the Nix file
for the Alpha machine by adding a `group` attribute set to "docker".
This change is essential to ensure that Traefik has the appropriate
permissions to interact with Docker containers. Proper configuration of
the group setting enhances the security and functionality of the
service, aligning with requirements to enable a seamless integration
with containerized applications.