Refactor Traefik configuration for clarity and robustness
This commit updates the `configuration.nix` file to improve the
structure and readability of the Traefik configuration. The changes
include a consolidation of settings into their appropriate nested
structure, enhancing both clarity and maintainability. Furthermore, the
`preStart` command for the `traefik` service was adjusted to check for
the existence of the `traefik-public` network before attempting to
create it. These updates not only help adhere to conventions but also
reduce potential errors during service startup.