Add Traefik dashboard and API configuration to Nix setup
This commit introduces modifications to the `configuration.nix` file
within the Traefik setup for the alpha machine. The changes enable the
Traefik dashboard and API access by adding an entry for the dashboard
and adjusting the dynamic configuration for Docker providers. This
update supports easier monitoring and management of services through the
Traefik interface, while also providing necessary options for debugging.
It's vital to note that the `insecure` option is set to true for
development purposes and should be disabled in production environments
to maintain security.