Add vim-surround and vim-repeat plugins to surround.nix
This commit introduces a new configuration file, `surround.nix`, for
integrating the `vim-surround` and `vim-repeat` plugins into the Vim
setup. The addition of these plugins enhances the functionality of the
Vim editor by allowing users to easily surround text with various
characters and repeat actions more efficiently. As part of this change,
updates have also been made to the Vim plugins environment, reinforcing
the commitment to improving the user experience for developers relying
on Vim. Additionally, the existing mappings in `mappings.vim` remain
unchanged, ensuring continuity in the editor's configuration.